refactor(bot): middleware to save the received interaction #5
