+++ bot/middlewares/index.js
@@ -1,0 +1,1 @@
+export { default as saveReceivedInteraction } from './save-received-interaction'
